# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */
export interface HealthStatus {
  status: string;
}

export interface BookSettings {
  /** @nullable */
  price: number | null;
  currency: string;
  vodafoneCash: string;
  instaPay: string;
  bookReady: boolean;
  cardReady: boolean;
  salesEnabled: boolean;
}

export type SettingsInputCurrency = typeof SettingsInputCurrency[keyof typeof SettingsInputCurrency];


export const SettingsInputCurrency = {
  egp: 'egp',
  usd: 'usd',
} as const;

export interface SettingsInput {
  /**
     * @maximum 100000
     * @exclusiveMinimum 0
     */
  price: number;
  currency: SettingsInputCurrency;
  /** @maxLength 100 */
  vodafoneCash: string;
  /** @maxLength 200 */
  instaPay: string;
  bookObjectPath?: string;
  salesEnabled: boolean;
}

export interface BookMe {
  signedIn: boolean;
  isAdmin: boolean;
}

export type BookOrderMethod = typeof BookOrderMethod[keyof typeof BookOrderMethod];


export const BookOrderMethod = {
  card: 'card',
  vodafone: 'vodafone',
  instapay: 'instapay',
} as const;

export type BookOrderStatus = typeof BookOrderStatus[keyof typeof BookOrderStatus];


export const BookOrderStatus = {
  pending: 'pending',
  paid: 'paid',
  rejected: 'rejected',
} as const;

export interface BookOrder {
  id: string;
  name: string;
  email: string;
  method: BookOrderMethod;
  status: BookOrderStatus;
  amount: number;
  currency: string;
  createdAt: string;
  hasReceipt: boolean;
  /** @nullable */
  checkoutUrl?: string | null;
  /** @nullable */
  reviewNote?: string | null;
}

export type OrderInputMethod = typeof OrderInputMethod[keyof typeof OrderInputMethod];


export const OrderInputMethod = {
  card: 'card',
  vodafone: 'vodafone',
  instapay: 'instapay',
} as const;

export type OrderInputLanguage = typeof OrderInputLanguage[keyof typeof OrderInputLanguage];


export const OrderInputLanguage = {
  ar: 'ar',
  en: 'en',
} as const;

export interface OrderInput {
  /**
     * @minLength 2
     * @maxLength 120
     */
  name: string;
  /** @maxLength 200 */
  email: string;
  method: OrderInputMethod;
  language: OrderInputLanguage;
  receiptObjectPath?: string;
}

export type UploadInputContentType = typeof UploadInputContentType[keyof typeof UploadInputContentType];


export const UploadInputContentType = {
  'image/jpeg': 'image/jpeg',
  'image/png': 'image/png',
  'application/pdf': 'application/pdf',
} as const;

export type UploadInputPurpose = typeof UploadInputPurpose[keyof typeof UploadInputPurpose];


export const UploadInputPurpose = {
  receipt: 'receipt',
  book: 'book',
} as const;

export interface UploadInput {
  /**
     * @minLength 1
     * @maxLength 200
     */
  name: string;
  /**
     * @minimum 1
     * @maximum 52428800
     */
  size: number;
  contentType: UploadInputContentType;
  purpose: UploadInputPurpose;
}

export interface UploadDestination {
  uploadURL: string;
  objectPath: string;
}

export type ReviewInputStatus = typeof ReviewInputStatus[keyof typeof ReviewInputStatus];


export const ReviewInputStatus = {
  paid: 'paid',
  rejected: 'rejected',
} as const;

export interface ReviewInput {
  status: ReviewInputStatus;
  /** @maxLength 1000 */
  reviewNote?: string;
}