# lib/api-zod/src/generated/types/bookSettings.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */

export interface BookSettings {
  /** @nullable */
  price: number | null;
  currency: string;
  vodafoneCash: string;
  instaPay: string;
  bookReady: boolean;
  cardReady: boolean;
  salesEnabled: boolean;
}