# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */
import type { OrderInputLanguage } from './orderInputLanguage';
import type { OrderInputMethod } from './orderInputMethod';

export interface OrderInput {
  /**
     * @minLength 2
     * @maxLength 120
     */
  name: string;
  /** @maxLength 200 */
  email: string;
  method: OrderInputMethod;
  language: OrderInputLanguage;
  receiptObjectPath?: string;
}