# lib/api-zod/src/generated/types/reviewInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * OpenAPI spec version: 1.0.0
 */
import type { ReviewInputStatus } from './reviewInputStatus';

export interface ReviewInput {
  status: ReviewInputStatus;
  /** @maxLength 1000 */
  reviewNote?: string;
}